requirements.txt
ccxt==4.2.45
pandas==2.0.3
numpy==1.24.3
requests
python-dotenv